+++ src/layer/vector/Renderer.js
@@ -1,0 +1,75 @@
+/*
+ * L.Renderer is a base class for renderer implementations (SVG, Canvas);
+ * handles renderer container, bounds and zoom animation.
+ */
+
+L.Renderer = L.Layer.extend({
+
+	options: {
+		// how much to extend the clip area around the map view (relative to its size)
+		// e.g. 0.1 would be 10% of map view in each direction; defaults to clip with the map view
+		padding: 0
+	},
+
+	initialize: function (options) {
+		L.setOptions(this, options);
+		L.stamp(this);
+	},
+
+	onAdd: function () {
+		if (!this._container) {
+			this._initContainer(); // defined by renderer implementations
+
+			if (this._zoomAnimated) {
+				L.DomUtil.addClass(this._container, 'leaflet-zoom-animated');
+			}
+		}
+
+		this.getPane().appendChild(this._container);
+		this._update();
+	},
+
+	onRemove: function () {
+		L.DomUtil.remove(this._container);
+	},
+
+	getEvents: function () {
+		var events = {
+			moveend: this._update
+		};
+		if (this._zoomAnimated) {
+			events.zoomanim = this._animateZoom;
+		}
+		return events;
+	},
+
+	_animateZoom: function (e) {
+		var scale = this._map.getZoomScale(e.zoom),
+		    offset = this._map._getCenterOffset(e.center)._multiplyBy(-scale)._add(this._bounds.min);
+
+		this._container.style[L.DomUtil.TRANSFORM] = L.DomUtil.getTranslateString(offset) + ' scale(' + scale + ') ';
+	},
+
+	_update: function () {
+		// update pixel bounds of renderer container (for positioning/sizing/clipping later)
+		var p = this.options.padding,
+		    size = this._map.getSize(),
+		    min = this._map.containerPointToLayerPoint(size.multiplyBy(-p)).round();
+
+		this._bounds = new L.Bounds(min, min.add(size.multiplyBy(1 + p * 2)).round());
+	}
+});
+
+
+L.Map.include({
+	// used by each vector layer to decide which renderer to use
+	getRenderer: function (layer) {
+		var renderer = layer.options.renderer || this.options.renderer ||
+				(L.SVG && L.SVG.instance) || (L.Canvas && L.Canvas.instance);
+
+		if (!this.hasLayer(renderer)) {
+			this.addLayer(renderer);
+		}
+		return renderer;
+	}
+});